backend/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6

idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
PyPDF2==3.0.1
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
pdf2image
PyMuPDF

httpx==0.27.0
flask==3.0.3
flask-cors==4.0.0
gtts==2.5.1
groq==0.9.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
numpy==1.26.4
torch==2.2.2
transformers==4.41.2